add z-index + split example
