+++ components/examples/window/src/example/index.ts
@@ -1,0 +1,5 @@
+export { default as example1 } from './example1';
+export { default as example2 } from './example2';
+export { default as example3 } from './example3';
+export { default as example4 } from './example4';
+export { default as example5 } from './example5';
